feat: remove Semaphore and use Mutex
Semaphore seems to work best when there are finite number of processes
in total, but once we're continiously spawning new tasks; it creates
more and more zombie processes that waits for 4-5 working ones. This
results in a horrendeous memory usage.

In the new method, I'm keeping track of the active number of tasks and
the id of the latest batch. And in the loop that spawns these tasks I
check if there are correct number of tasks at that moment.

It has its flaws, like in the case where I have to check once more in the
thread that if I'm not exceeding MAX_TASK_COUNT, but it works so far.